student_management_project/app/attendance/serializers.py
from rest_framework import serializers
from .models import Attendance, AttendanceReport

class AttendanceSerializer(serializers.ModelSerializer):
    class Meta:
        model = Attendance
        fields = '__all__'

class AttendanceReportSerializer(serializers.ModelSerializer):
    class Meta:
        model = AttendanceReport
        fields = '__all__'