student_management_project/requirements.txt
django>=5.0.1
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
djangorestframework-simplejwt>=5.3.1
PyJWT>=2.8.0
channels>=4.0.0
channels-redis>=4.1.1
daphne>=4.0.0
psycopg2-binary>=2.9.9
redis>=5.0.1
celery>=5.3.6
Pillow>=10.2.0
python-dotenv>=1.0.1
gunicorn>=21.2.0